Fix rare-ish camera jump when arriving at a flight destination (#1149)
* Fix incorrect height at end of flight.

In the Sublevels Sample, sometimes, upon arrival in Paris, the flight interpolation percentage would be exactly 1.0. Not just "displays as 1.0", but exactly equal to 1.0. And when that happened, the interpolation of the altitude flight curve would return 1.0, so the flight would put the altitude at its maximum rather than its destination height. This makes absolutely no sense, because the altitude interpolation key points are (0.0, 0.0), (0.5, 1.0), (1.0, 0.0). In other words, it should be 0.0 (minimum height) at the start and end, and 1.0 (maximum height) in the middle, just what you'd expect for altitude on a camera flight. Yet asking for the interpolated value at exactly 1.0 would return that value at 0.5 instead of the value at 1.0.

So stepping into that curve interpolation, I found it's doing some fancy cubic interpolation with tangents and whatnot. Ok, fine, but I'm literally asking for one of the keypoints! Well apparently the implementation of this cubic interpolation solves some cubic equation and comes up with three roots. It then picks the one that is between 0.0 and 1.0. Three of the roots are well outside that range, but one is very very very slightly larger than 1.0. So Unreal decides none of these solutions are correct and decides to just make up an answer instead, 0.0.

This commit works around this problem by ensuring that we never
interpolate at t=1.0. Instead, t=1.0 is simply treated as being at the
end of the flight.

* Update CHANGES.md.

* Add a test for altitude spike fix.

* does -> should

* Formatting.

* Don't attempt to initiate a flight if there's no Controller.

+++ Source/CesiumRuntime/Public/GlobeAwareDefaultPawn.h
@@ -227,7 +227,7 @@ private:
   void _moveAlongViewAxis(EAxis::Type axis, double Val);
   void _moveAlongVector(const FVector& axis, double Val);
   void _interruptFlight();
-  void _interpolateFlightPosition(double percentage, glm::dvec3& out) const;
+  void _interpolateFlightPosition(float percentage, glm::dvec3& out) const;
 
   /**
    * @brief Advance the camera flight based on the given time delta.
@@ -248,7 +248,7 @@ private:
   // helper variables for FlyToLocation
   bool _bFlyingToLocation = false;
   bool _bCanInterruptFlight = false;
-  double _currentFlyTime = 0.0;
+  float _currentFlyTime = 0.0f;
 
   double _flyToSourceAltitude = 0.0;
   double _flyToDestinationAltitude = 0.0;
